% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_MLnet.R
\name{DrawMLnet}
\alias{DrawMLnet}
\title{Save and Draw the Multi-layer Signaling Network}
\usage{
DrawMLnet(MLnetList, LigClu, RecClu, outdir = NULL, PyHome, plotMLnet = F)
}
\arguments{
\item{MLnetList}{The output of Run_MLnet function.}

\item{LigClu}{character: The neighboring cell}

\item{RecClu}{character: The central cell}

\item{outdir}{The name of output folder}

\item{PyHome}{The path to PYTHONHOME. The PyHome parameter is invalidated if only the plotMLnet parameter is set to TRUE.}

\item{plotMLnet}{Whether to draw the multi-layer signaling network. The default setting is FALSE.}
}
\description{
This  function save the results generated by Rub_scMLnet function and visualize the results by pymnet package (python).
The generated multi-layer signaling network is saved in the folder named by the LigClu, RecClu and outdir parameters.
}
