% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_scMLnet.R
\name{RunMLnet}
\alias{RunMLnet}
\title{Title Generate Multi-layer Signal Networks}
\usage{
RunMLnet(
  GCMat,
  BarCluFile,
  RecClu,
  LigClu,
  pval = 0.05,
  logfc = 0.15,
  LigRecLib = "./database/LigRec.txt",
  TFTarLib = "./database/TFTargetGene.txt",
  RecTFLib = "./database/RecTF.txt"
)
}
\arguments{
\item{GCMat}{scRNA-seq data. The gene expression matrix（raw) with rows as genes (gene symbols) and columns as cells.}

\item{BarCluFile}{The annotation results for clustering. The first column is barcode and the second is cell type.}

\item{RecClu}{character: The central cell}

\item{LigClu}{character: The neighboring cell}

\item{pval}{Screening threshold for getting high expressed gene in cluster. The default setting is 0.05.}

\item{logfc}{Screening threshold for getting high expressed gene in cluster. The default setting is 0.15.}

\item{LigRecLib}{The file path of Ligand-Receptor interactions. The default setting is 'LigRec.txt' file in the /databases/ folder.}

\item{TFTarLib}{The file path of TF-TarFget gene interactions. The default setting is 'TFTargetGene.txt' file in the /databases/ folder.}

\item{RecTFLib}{The file path of Receptor-TF interactions. The default setting is 'RecTF.txt' file in the /databases/ folder.}
}
\value{
A list consists of the Ligand_Receptor, Receptor_TF and TF_TarGene signaling subnetwork.
The signaling subnetwork is returned as a dataframe object, including three columns:
the first column is molecule A, and the second column is molecule B.
There is an interaction between the molecules A and B, which correspond to the ligand, receptor,
transcription factor or target gene. The third column is used to visualize the signaling subnetwork.
}
\description{
This function constructs the Ligand_Receptor, Receptor_TF, TF_TarGene networks between the central cell and neighboring cells according to scRNA-Seq expression matrix and barcode table
}
